refactor: 서비스 레이어 interface 를 iface 패키지로 분리
