+++ src/main/java/com/equip/equiprental/board/service/BoardServiceImpl.java
@@ -5,13 +5,14 @@ import com.equip.equiprental.board.domain.BoardStatus;
 import com.equip.equiprental.board.domain.BoardType;
 import com.equip.equiprental.board.dto.*;
 import com.equip.equiprental.board.repository.BoardRepository;
+import com.equip.equiprental.board.service.iface.BoardService;
 import com.equip.equiprental.common.dto.PageResponseDto;
 import com.equip.equiprental.common.dto.SearchParamDto;
 import com.equip.equiprental.common.exception.CustomException;
 import com.equip.equiprental.common.exception.ErrorType;
 import com.equip.equiprental.filestorage.domain.FileMeta;
 import com.equip.equiprental.filestorage.repository.FileRepository;
-import com.equip.equiprental.filestorage.service.FileService;
+import com.equip.equiprental.filestorage.service.iface.FileService;
 import com.equip.equiprental.member.domain.Member;
 import com.equip.equiprental.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
