+++ src/main/java/com/equip/equiprental/equipment/service/CategoryServiceImpl.java
@@ -6,6 +6,7 @@ import com.equip.equiprental.equipment.dto.CategoryDto;
 import com.equip.equiprental.equipment.dto.SubCategoryDto;
 import com.equip.equiprental.equipment.repository.CategoryRepository;
 import com.equip.equiprental.equipment.repository.SubCategoryRepository;
+import com.equip.equiprental.equipment.service.iface.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +14,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class CategoryServiceImpl implements CategoryService{
+public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final SubCategoryRepository subCategoryRepository;
