+++ src/main/java/com/equip/equiprental/equipment/service/EquipmentServiceImpl.java
@@ -13,9 +13,10 @@ import com.equip.equiprental.equipment.dto.*;
 import com.equip.equiprental.equipment.repository.EquipmentItemRepository;
 import com.equip.equiprental.equipment.repository.EquipmentRepository;
 import com.equip.equiprental.equipment.repository.SubCategoryRepository;
+import com.equip.equiprental.equipment.service.iface.EquipmentService;
 import com.equip.equiprental.filestorage.domain.FileMeta;
 import com.equip.equiprental.filestorage.repository.FileRepository;
-import com.equip.equiprental.filestorage.service.FileService;
+import com.equip.equiprental.filestorage.service.iface.FileService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
