+++ src/main/java/com/equip/equiprental/member/controller/MemberController.java
@@ -6,7 +6,7 @@ import com.equip.equiprental.common.dto.ResponseDto;
 import com.equip.equiprental.common.dto.SearchParamDto;
 import com.equip.equiprental.member.dto.*;
 import com.equip.equiprental.common.interceptor.RequestTraceIdInterceptor;
-import com.equip.equiprental.member.service.MemberService;
+import com.equip.equiprental.member.service.iface.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
