+++ src/main/java/com/equip/equiprental/member/service/DepartmentServiceImpl.java
@@ -3,6 +3,7 @@ package com.equip.equiprental.member.service;
 import com.equip.equiprental.member.domain.Department;
 import com.equip.equiprental.member.dto.DepartmentDto;
 import com.equip.equiprental.member.repository.DepartmentRepository;
+import com.equip.equiprental.member.service.iface.DepartmentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +11,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class DepartmentServiceImpl implements DepartmentService{
+public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
 
     @Override
