+++ src/main/java/com/equip/equiprental/rental/controller/RentalItemController.java
@@ -8,8 +8,7 @@ import com.equip.equiprental.common.dto.SearchParamDto;
 import com.equip.equiprental.common.interceptor.RequestTraceIdInterceptor;
 import com.equip.equiprental.rental.dto.AdminRentalItemDto;
 import com.equip.equiprental.rental.dto.ExtendRentalItemDto;
-import com.equip.equiprental.rental.dto.UserRentalItemDto;
-import com.equip.equiprental.rental.service.RentalItemService;
+import com.equip.equiprental.rental.service.iface.RentalItemService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
