+++ src/main/java/com/equip/equiprental/rental/service/iface/RentalItemService.java
@@ -1,10 +1,9 @@
-package com.equip.equiprental.rental.service;
+package com.equip.equiprental.rental.service.iface;
 
 import com.equip.equiprental.common.dto.PageResponseDto;
 import com.equip.equiprental.common.dto.SearchParamDto;
 import com.equip.equiprental.rental.dto.AdminRentalItemDto;
 import com.equip.equiprental.rental.dto.ExtendRentalItemDto;
-import com.equip.equiprental.rental.dto.UserRentalItemDto;
 
 public interface RentalItemService {
     PageResponseDto<AdminRentalItemDto> getAdminRentalItemLists(SearchParamDto paramDto);
